reorganize matrix code
