+++ src/votekit/matrices/candidate/__init__.py
@@ -1,0 +1,3 @@
+from .comentions import comention, comention_above, comentions_matrix  # noqa
+from .boost import boost_prob, boost_matrix  # noqa
+from .candidate_distance import candidate_distance, candidate_distance_matrix  # noqa
