+++ src/votekit/matrices/candidate/candidate_distance.py
@@ -1,8 +1,8 @@
-from ..ballot import Ballot
-from ..pref_profile import PreferenceProfile
+from ...ballot import Ballot
+from ...pref_profile import PreferenceProfile
 import numpy as np
 import itertools as it
-from ._utils import _convert_dict_to_matrix
+from .._utils import _convert_dict_to_matrix
 
 
 def candidate_distance(i: str, j: str, ballot: Ballot) -> float:
