+++ src/votekit/matrices/candidate/comentions.py
@@ -1,6 +1,6 @@
-from ..ballot import Ballot
-from ..pref_profile import PreferenceProfile
-from ._utils import _convert_dict_to_matrix
+from ...ballot import Ballot
+from ...pref_profile import PreferenceProfile
+from .._utils import _convert_dict_to_matrix
 import numpy as np
 from fractions import Fraction
 from typing import Union
