+++ tests/matrices/candidate/boost/test_boost_prob.py
